==> your-project/02. Further web scrapping trials.ipynb <==
# 2017 happiness report

##### import all libraries needed

import requests
import pandas as pd
from bs4 import BeautifulSoup
import os


#### call url and pass through "soup"

url = "https://en.wikipedia.org/wiki/World_Happiness_Report"
html = requests.get(url).content
soup = BeautifulSoup(html, "lxml")

# look for the table class on the website developer tools
table = soup.find_all('table',{'class':"mw-collapsible"})[5]


#### check on the rows and layout of the table
!!! failed as this is an embedded list into a collapsible sub-section - data could not be extracted!!!

rows = table.find_all('tr')

# to see the overall layout of the table use:
#rows = [row.text.strip().split("\n") for row in rows]
#rows
# edit out all info we do not need to get "clean data"
rows = [row.text.replace("\n\n", ",").replace("\n", ",").replace("\xa0", "").strip(",").split(",") for row in rows]
#rows [:2]

#### create table

"""colnames = rows[1]
data = rows[2:]

happiness_2017_extended = pd.DataFrame(data, columns=colnames)
happiness_2017_extended.head(20)
"""

# Austrial election results 

### 2017 elections

url = "https://wahl17.bmi.gv.at/"
html = requests.get(url).content
soup = BeautifulSoup(html, "lxml")
#print(soup.prettify())

rows = soup.find_all('tr')
"""rows = [row.text.strip().split("\n") for row in rows]
rows
"""

# as we are only looking for the votes ONLY for the extreme parties
# we only need to select the FPÖ row (index = 3) and the KPÖ (not on list as percetnage too little)
FPO_2017 = [row.text.strip().split("\n") for row in rows][3]
FPO_2017

FPO_2017 = [row.replace(",0\xa0%51", "").replace("FPÖ", "") for row in FPO_2017][0]
FPO_2017 = int(FPO_2017)
FPO_2017

FPÖ: 26.00%

We realised that this made no sense as it is not a atuomatable functin that we can use for more examples, and therefore we did not continue with this method

## 2013 elections

url = "http://wahl13.bmi.gv.at/"
html = requests.get(url).content
soup = BeautifulSoup(html, "lxml")
#print(soup.prettify())

table = soup.find_all('section',{'class':'content-box table'})[0]
rows = table.find_all('tr')
rows = [row.text.replace("\n\n", ",").replace("\n", ",").replace("\xa0", "").strip(",").split(",") for row in rows]

FPO_2013 = rows[11][0]+rows[11][1]
FPO_2013

FPO_2013 = FPO_2013.replace("FPÖ962.313", "").replace("5%857.02917", "")
FPO_2013
