==> your-project/03. Database_creation_seeding.ipynb <==
### Creating connection with the project's database and creating the tables in it

from sqlalchemy import create_engine
import pymysql
import pandas as pd
import numpy as np

driver = "mysql+pymysql:"
user = "root"
password = "ironhack"
ip = "35.188.185.234"
database = "Project_3"

connection_string = f"{driver}//{user}:{password}@{ip}/{database}"
engine = create_engine(connection_string)

query_country = """
CREATE TABLE country(
id INT(17) NOT NULL AUTO_INCREMENT,
country VARCHAR(25),

PRIMARY KEY(id)
);
"""


engine.execute(query_country)

query_elections = """
CREATE TABLE elections(
id INT(17) NOT NULL AUTO_INCREMENT,
year INT,
id_country INT(17),
proportion_votes FLOAT,
extreme_type VARCHAR(30),

PRIMARY KEY(id),
FOREIGN KEY(id_country) REFERENCES country(id)
);
"""



engine.execute(query_elections)

happiness_all = pd.read_csv("../Data/happiness_all.csv")

happiness_all.to_sql("happiness", con=engine, if_exists="append", index=False)

#Below couple of cells were used at the beginning, but in the end the happiness final dataframe, had many 
#more columns so we decided to get rid of the rigid table structure, and created the table in another way (above).

query_happiness = """
CREATE TABLE happiness(
id INT NOT NULL AUTO_INCREMENT,
id_year INT(17),
id_country INT(17),
score FLOAT,

PRIMARY KEY(id),
FOREIGN KEY(id_country) REFERENCES country(id)
);
"""


engine.execute(query_happiness)

### Checking the tables created

query_tables = """
SHOW TABLES
"""

db_tables = pd.read_sql(query_tables, engine)
db_tables

### Inputting data into our database and checking operation success

countries = pd.read_csv("../Data/countries.csv")

countries

countries.to_sql("country", con=engine, if_exists="append", index=False)

engine.execute("SELECT * FROM country").fetchall()

elections = pd.read_csv("../Data/Elections_10_countries.csv")

elections.head()

elections.to_sql("elections", con=engine, if_exists="append", index=False)

# To check whether all data has been input correctly into our database
engine.execute("SELECT * FROM elections").fetchall()

engine.execute("SELECT * FROM happiness").fetchall()

#### Dropping innecessary tables

engine.execute("DROP TABLE happiness")
#engine.execute("DROP TABLE happiness_all")

engine.execute("DROP TABLE elections")
