==> your-project/05. Data Cleaning.ipynb <==
import pandas as pd
import numpy as np

##  Countries of analysis:

country_list = ["Spain", "Netherlands", "France", "Greece", "Austria", "Denmark", "Belgium", "Italy", "Poland", "Sweden"]

List of headers in order to match to the SQL database (reference)

header_list = ["Year", "Country", "Score", "GDP", "Social support", 
               "Healthy life expectancy", "Freedome", "Corruption"]

country_id = [key+1 for key in range(10)]

country_id

countries = pd.DataFrame({'country': country_list, 'country_id': country_id})

countries.head()

## Cleaning of Happiness 2017

#### open file

happiness_2017 = pd.read_csv("../Data/backup_raw_data/happiness_2017_extended.csv")

#### select countries

happiness_2017 = happiness_2017[(happiness_2017["Country"].isin(country_list))]
happiness_2017.columns

##### change the col names + drop empty cols

happiness_2017 = happiness_2017.dropna(axis=1)
happiness_2017 = happiness_2017.rename(columns={"Happiness score":"Score", "GDP per capita":"GDP",
                                               "Freedom to make life choices":"Freedom"})


happiness_2017

#### save to new csv

happiness_2017.to_csv("../Data/happiness_2017.csv")

# Cleaning of happiness 2016
code steps explained in previous section, same process used for all tables

# read
happiness_2016 = pd.read_csv("../Data/backup_raw_data/happiness_2016_extended.csv")

# check for countries 
happiness_2016 = happiness_2016[(happiness_2016["Country"].isin(country_list))]

# rename columns for consistencey to join
happiness_2016 = happiness_2016.rename(columns={"Happiness score":"Score", "GDP per capita":"GDP", 
                                                "Freedom to make life choices":"Freedom"})

# frop Na column 
happiness_2016 = happiness_2016.dropna(axis=1)

happiness_2016


# save
happiness_2016.to_csv("../Data/happiness_2016.csv")

# Cleaning of happines other years

# read file
happiness_older = pd.read_csv("../Data/backup_raw_data/happiness_2015_to_2010_extended.csv")

# select countries
happiness_older = happiness_older[(happiness_older["country"].isin(country_list))]

# make column names consistent to join 
happiness_older = happiness_older.rename(columns={"Life Ladder":"Score", "Log GDP per capita":"GDP", 
                                                  "Healthy life expectancy at birth":"Healthy life expectancy", 
                                                  "Freedom to make life choices":"Freedom"})

# two columns with NaN as csv file had a cell merge
happiness_older = happiness_older.dropna(axis=1)

# maintain all columns consistent with all files of Happiness. 
# 2015 onwards dont't have pos, neg affect and confidence in national government 
happiness_older = happiness_older.drop(["Positive affect"], axis=1)
happiness_older = happiness_older.drop(["Negative affect"], axis=1)

happiness_older.head()

# save 
happiness_older.to_csv("../Data/happiness_2015_and_older.csv")

# Cleaning final data to insert it into the database

## Cleaning countries dataframe

countries = countries.rename(columns={"country_id":"id"})

countries_columns = ["id", "country"]

countries = countries[countries_columns]

countries.to_csv("../Data/countries.csv", index=False)

## Cleaning elections data

elections = pd.read_csv("../Data/backup_raw_data/Elections_10_countries_raw.csv")

elections.head()

#### Renaming the columns to the way we have defined in the database

elections = elections.rename(columns={"Country":"id_country", "Year":"year", "Proportion":"proportion_votes", "Type":"extreme_type"})

#### Cleaning misleading names

elections["id_country"] = elections["id_country"].str.replace("French", "France")

#### Changing name of countries per ids, since this is consistent with the way we designed the database

country_id_dict = dict(zip(country_id, country_list))

country_id_dict

elections = elections.replace({"id_country": country_id_dict})

#elections.head()

#### Changing the order of columns

elections.columns

elections_columns = ["year", "id_country", "proportion_votes", "extreme_type"]

elections = elections[elections_columns]

elections.head()

elections.to_csv("../Data/Elections_10_countries.csv", index=False)

## Adapting happiness reports
As we relised that part of the measurement scale of the reports did not match after the merge, we took the merged file from 2019 provided by the happiness report website and used that to analyse the correlation of individual variables to the election rates of chosen countries.

(see happines_cons - standing for consistent variable mesurements)

#happiness = pd.read_csv("../Data/backup_raw_data/happiness_all_raw.csv")
happiness_cons = pd.read_csv("../Data/backup_raw_data/happiness_all_raw_consistent_eval.csv")

#happiness.head()
happiness_cons.head()

#### Renaming column so that we can make a merge between happiness and countries dataframes

#happiness = happiness.rename(columns={"Country":"country"})
happiness_cons = happiness_cons.rename(columns={"Country name":"country", "Life Ladder":"Score", 
                                                "Log GDP per capita":"GDP", 
                                                "Healthy life expectancy at birth":"Healthy life expectancy",
                                                "Freedom to make life choices":"Freedom"})

#happiness.head()
happiness_cons.head()

### Merging dataframes to then be able to convert countries per ids more easily

#happiness = pd.merge(happiness,countries,on="country") 
happiness_cons = pd.merge(happiness_cons,countries,on="country") 

#happiness.columns
happiness_cons.columns

### Ordering columns

#happiness_columns = ["id", "country", "Year", "Score", "GDP", "Social support", "Healthy life expectancy", "Freedom",
                     #"Generosity", "Perceptions of corruption", "Confidence in national government"]
       

#happiness = happiness[happiness_columns]

happiness_columns = ["id", "country", "Year", "Score", "GDP", "Social support", "Healthy life expectancy", 
                     "Freedom", "Generosity", "Perceptions of corruption", "Confidence in national government"]
       

happiness_cons = happiness_cons[happiness_columns]

#happiness_columns = list(happiness.columns)
happiness_cons_columns = list(happiness_cons.columns)

#happiness_columns
happiness_cons_columns

### Replacing names for consistency between dataframes and database. 
Used a function so it can be used more times easily

def lower_case(list):
    new_list = []
    for word in list:
        new_word = word.lower()
        new_list.append(new_word)
    return new_list


# happiness_columns = lower_case(happiness_columns)
happiness_cons_columns = lower_case(happiness_columns)

print(happiness_columns)
print (happiness_cons_columns)

#happiness.head()
happiness_cons.head()

#happiness = happiness.rename(columns={"id":"id_country"})
happiness_cons = happiness_cons.rename(columns={"id":"id_country"})

### Dropping country names, since we have merged with countries and we have now country ids

#happiness = happiness.drop("country",axis=1)
happiness_cons = happiness_cons.drop("country",axis=1)

#happiness.head()
happiness_cons.head()

#happiness.to_csv("../Data/happiness_all.csv", index=False)
happiness_cons.to_csv("../Data/happiness_all_constant.csv", index=False)
