==> your-project/06. Combined_happiness_index_all_year.ipynb <==
import pandas as pd
import numpy as np


ha_2015_older = pd.read_csv("happiness_2015_and_older.csv")

ha_2015_older = ha_2015_older.drop(columns=["Unnamed: 0"])

ha_2015_older.head()

ha_2016 = pd.read_csv("happiness_2016.csv")
ha_2016 = ha_2016.drop(columns=["Unnamed: 0"])
ha_2016["Year"] = 2016

ha_2017 = pd.read_csv("happiness_2017.csv")
ha_2017 = ha_2017.drop(columns=["Unnamed: 0"])
ha_2017 = ha_2017.rename(columns={" Generosity":"Generosity"})
ha_2017["Year"] = 2017

ha_2018 = pd.read_csv("happiness_2018.csv")
ha_2018 = ha_2018.drop(columns=["Unnamed: 0"])
ha_2018 = ha_2018.rename(columns={"Generosity importance":"Generosity"})
ha_2018["Year"] = 2018

ha_2016_later = pd.concat([ha_2016, ha_2017, ha_2018], axis=0, sort=False)
ha_2016_later = ha_2016_later.reset_index(drop=True)

ha_2016_later = ha_2016_later[["Country", "Year", "Score", "GDP", "Social support", "Healthy life expectancy", "Freedom", "Generosity", "Perceptions of corruption"]]
ha_2016_later["Confidence in national government"] = np.nan
ha_2016_later.head()

ha_2015_older = ha_2015_older.rename(columns = {"year":"Year", "country":"Country"})
ha_2015_older["Generosity"] = np.nan
ha_2015_older["Perceptions of corruption"] = np.nan
ha_2015_older = ha_2015_older[["Country", "Year", "Score", "GDP", "Social support", "Healthy life expectancy", "Freedom", "Generosity", "Perceptions of corruption", "Confidence in national government"]]
ha_2015_older.head()

ha_all = pd.concat([ha_2015_older, ha_2016_later])

ha_all = ha_all.reset_index(drop=True)
ha_all.head()
ha_all.to_csv("./backup_raw_data/happiness_all_raw.csv", index=False)
