==> your-project/07. Project_3_correlation.ipynb <==
import pandas as pd
import numpy as np

df_countries = pd.read_csv("/home/chao/Desktop/Ironhack/Projects/Project-Week-3-Data-Thieves/Data/countries.csv")
df_happiness = pd.read_csv("/home/chao/Desktop/Ironhack/Projects/Project-Week-3-Data-Thieves/Data/happiness_all.csv")
df_elections = pd.read_csv("/home/chao/Desktop/Ironhack/Projects/Project-Week-3-Data-Thieves/Data/Elections_10_countries.csv")
df_happiness = df_happiness[df_happiness["Year"] > 2010]

happiness = pd.merge(df_happiness, df_countries, how ="left", left_on='id_country', right_on='id')
happiness = happiness.rename(columns = {"Year":"year"})
del happiness["id"]

austria = happiness[happiness["country"] == "Austria"]
austria = austria.merge(df_elections, how="left", on=["id_country", "year"])
del austria["id_country"]
austria = austria.groupby("year").sum()
austria = austria.drop([2011, 2015]).reset_index()
austria = austria.replace(0,np.nan)
austria["proportion_votes"] = austria["proportion_votes"].bfill().ffill()
austria_corr = austria.corr(method="pearson")
austria_corr
#austria_result = austria_corr["proportion_votes"]

belgium = happiness[happiness["country"] == "Belgium"]
belgium = belgium.merge(df_elections, how="left", on=["id_country", "year"])
del belgium["id_country"]
belgium = belgium.groupby("year").sum()
belgium = belgium.drop([2011, 2012, 2016, 2017]).reset_index()
belgium = belgium.replace(0,np.nan)
belgium["proportion_votes"] = belgium["proportion_votes"].bfill().ffill()
belgium_corr = belgium.corr(method="pearson")
belgium_corr
#belgium_result = belgium_corr["proportion_votes

denmark = happiness[happiness["country"] == "Denmark"]
denmark = denmark.merge(df_elections, how="left", on=["id_country", "year"])
del denmark["id_country"]
denmark = denmark.groupby("year").sum()
denmark = denmark.drop([2011, 2012, 2013, 2017]).reset_index()
denmark = denmark.replace(0,np.nan)
denmark["proportion_votes"] = denmark["proportion_votes"].bfill().ffill()
denmark_corr = denmark.corr(method="pearson")
denmark_corr
#denmark_result = denmark_corr["proportion_votes

france = happiness[happiness["country"] == "France"]
france = france.merge(df_elections, how="left", on=["id_country", "year"])
del france["id_country"]
france = france.groupby("year").sum()
france = france.drop([2011, 2012, 2016, 2017]).reset_index()
france = france.replace(0,np.nan)
france["proportion_votes"] = france["proportion_votes"].bfill().ffill()
france_corr = france.corr(method="pearson")
france_corr
#france_result = france_corr["proportion_votes

greece = happiness[happiness["country"] == "Greece"]
greece = greece.merge(df_elections, how="left", on=["id_country", "year"])
del greece["id_country"]
greece = greece.groupby("year").sum()
greece = greece.drop([2011, 2012, 2013, 2017]).reset_index()
greece = greece.replace(0,np.nan)
greece["proportion_votes"] = greece["proportion_votes"].bfill().ffill()
greece_corr = greece.corr(method="pearson")
greece_corr
#greece_result = greece_corr["proportion_votes

italy = happiness[happiness["country"] == "Italy"]
italy = italy.merge(df_elections, how="left", on=["id_country", "year"])
del italy["id_country"]
italy = italy.groupby("year").sum()
italy = italy.drop([2014, 2015, 2016]).reset_index()
italy = italy.replace(0,np.nan)
italy["proportion_votes"] = italy["proportion_votes"].bfill().ffill()
italy_corr = italy.corr(method="pearson")
italy_corr
#italy_result = italy_corr["proportion_votes

netherlands = happiness[happiness["country"] == "Netherlands"]
netherlands = netherlands.merge(df_elections, how="left", on=["id_country", "year"])
del netherlands["id_country"]
netherlands = netherlands.groupby("year").sum()
netherlands = netherlands.drop([2014, 2015]).reset_index()
netherlands = netherlands.replace(0,np.nan)
netherlands["proportion_votes"] = netherlands["proportion_votes"].bfill().ffill()
netherlands_corr = netherlands.corr(method="pearson")
netherlands_corr
#netherlands_result = netherlands_corr["proportion_votes

poland = happiness[happiness["country"] == "Poland"]
poland = poland.merge(df_elections, how="left", on=["id_country", "year"])
del poland["id_country"]
poland = poland.groupby("year").sum()
poland = poland.drop([2017, 2018]).reset_index()
poland = poland.replace(0,np.nan)
poland["proportion_votes"] = poland["proportion_votes"].bfill().ffill()
poland_corr = poland.corr(method="pearson")
poland_corr
#poland_result = poland_corr["proportion_votes

spain = happiness[happiness["country"] == "Spain"]
spain = spain.merge(df_elections, how="left", on=["id_country", "year"])
del spain["id_country"]
spain = spain.groupby("year").sum()
spain = spain.drop([2013, 2014]).reset_index()
spain = spain.replace(0,np.nan)
spain["proportion_votes"] = spain["proportion_votes"].bfill().ffill()
spain_corr = spain.corr(method="pearson")
spain_corr
#spain_result = spain_corr["proportion_votes

spain

sweden = happiness[happiness["country"] == "Sweden"]
sweden = sweden.merge(df_elections, how="left", on=["id_country", "year"])
del sweden["id_country"]
sweden = sweden.groupby("year").sum()
sweden = sweden.drop([2011, 2012, 2016]).reset_index()
sweden = sweden.replace(0,np.nan)
sweden["proportion_votes"] = sweden["proportion_votes"].bfill().ffill()
sweden_corr = sweden.corr(method="pearson")
sweden_corr
#sweden_result = sweden_corr["proportion_votes
